split navigations
